Crate tracing_opentelemetry

Tracing OpenTelemetry

An opentelemetry layer for the tracing library.

#[macro_use]
extern crate tracing;

use opentelemetry::{api::Provider, sdk};
use tracing_opentelemetry::OpentelemetryLayer;
use tracing_subscriber::{Layer, Registry};

fn main() {
    // Create a new tracer
    let tracer = sdk::Provider::default().get_tracer("service_name");

    // Create a new tracing layer
    let layer = OpentelemetryLayer::with_tracer(tracer);

    let subscriber = layer.with_subscriber(Registry::default());

    // Trace executed code
    tracing::subscriber::with_default(subscriber, || {
        let root = span!(tracing::Level::TRACE, "app_start", work_units = 2);
        let _enter = root.enter();

        error!("This event will be logged in the root span.");
    });
}

Structs

OpentelemetryLayer
OpenTelemetry layer for use in a project that uses tracing.