tracing_opentelemetry

Struct OpenTelemetryLayer

pub struct OpenTelemetryLayer<S, T> { /* private fields */ }

An OpenTelemetry propagation layer for use in a project that uses tracing.

Implementations

impl<S, T> OpenTelemetryLayer<S, T>
where S: Subscriber + for<'span> LookupSpan<'span>, T: Tracer + PreSampledTracer + 'static,

pub fn new(tracer: T) -> Self

Set the Tracer that this layer will use to produce and track OpenTelemetry Spans.

Examples

use tracing_opentelemetry::OpenTelemetryLayer;
use tracing_subscriber::layer::SubscriberExt;
use opentelemetry::trace::TracerProvider as _;
use tracing_subscriber::Registry;

// Create an OTLP pipeline exporter for a `trace_demo` service.

let otlp_exporter = opentelemetry_otlp::SpanExporter::builder()
    .with_tonic()
    .build()
    .unwrap();

let tracer = opentelemetry_sdk::trace::SdkTracerProvider::builder()
    .with_simple_exporter(otlp_exporter)
    .build()
    .tracer("trace_demo");

// Create a layer with the configured tracer
let otel_layer = OpenTelemetryLayer::new(tracer);

// Use the tracing subscriber `Registry`, or any other subscriber
// that impls `LookupSpan`
let subscriber = Registry::default().with(otel_layer);

pub fn with_tracer<Tracer>( self, tracer: Tracer, ) -> OpenTelemetryLayer<S, Tracer>
where Tracer: Tracer + PreSampledTracer + 'static,

Set the Tracer that this layer will use to produce and track OpenTelemetry Spans.

Examples

use tracing_subscriber::layer::SubscriberExt;
use tracing_subscriber::Registry;
use opentelemetry::trace::TracerProvider;

// Create an OTLP pipeline exporter for a `trace_demo` service.

let otlp_exporter = opentelemetry_otlp::SpanExporter::builder()
    .with_tonic()
    .build()
    .unwrap();

let tracer = opentelemetry_sdk::trace::SdkTracerProvider::builder()
    .with_simple_exporter(otlp_exporter)
    .build()
    .tracer("trace_demo");

// Create a layer with the configured tracer
let otel_layer = tracing_opentelemetry::layer().with_tracer(tracer);

// Use the tracing subscriber `Registry`, or any other subscriber
// that impls `LookupSpan`
let subscriber = Registry::default().with(otel_layer);

pub fn with_error_fields_to_exceptions( self, error_fields_to_exceptions: bool, ) -> Self

Sets whether or not span and event metadata should include OpenTelemetry exception fields such as exception.message and exception.backtrace when an Error value is recorded. If multiple error values are recorded on the same span/event, only the most recently recorded error value will show up under these fields.

These attributes follow the OpenTelemetry semantic conventions for exceptions.

By default, these attributes are recorded. Note that this only works for (dyn Error + 'static). See Implementations on Foreign Types of tracing::Value or OpenTelemetryLayer::with_error_events_to_exceptions

pub fn with_error_events_to_status(self, error_events_to_status: bool) -> Self

Sets whether or not an event considered for exception mapping (see [OpenTelemetryLayer::with_error_recording]) should be propagated to the span status error description.

By default, these events do set the span status error description.

pub fn with_error_events_to_exceptions( self, error_events_to_exceptions: bool, ) -> Self

Sets whether or not a subset of events following the described schema are mapped to events following the OpenTelemetry semantic conventions for exceptions.

Only events without a message field (unnamed events) and at least one field with the name error are considered for mapping.

By default, these events are mapped.

pub fn with_error_records_to_exceptions( self, error_records_to_exceptions: bool, ) -> Self

Sets whether or not reporting an Error value on an event will propagate the OpenTelemetry exception fields such as exception.message and exception.backtrace to the corresponding span. You do not need to enable with_exception_fields in order to enable this. If multiple error values are recorded on the same span/event, only the most recently recorded error value will show up under these fields.

These attributes follow the OpenTelemetry semantic conventions for exceptions.

By default, these attributes are propagated to the span. Note that this only works for (dyn Error + 'static). See Implementations on Foreign Types of tracing::Value or OpenTelemetryLayer::with_error_events_to_exceptions

pub fn with_location(self, location: bool) -> Self

Sets whether or not span and event metadata should include OpenTelemetry attributes with location information, such as the file, module and line number.

These attributes follow the OpenTelemetry semantic conventions for source locations.

By default, locations are enabled.

pub fn with_tracked_inactivity(self, tracked_inactivity: bool) -> Self

Sets whether or not spans metadata should include the busy time (total time for which it was entered), and idle time (total time the span existed but was not entered).

By default, inactivity tracking is enabled.

pub fn with_threads(self, threads: bool) -> Self

Sets whether or not spans record additional attributes for the thread name and thread ID of the thread they were created on, following the OpenTelemetry semantic conventions for threads.

By default, thread attributes are enabled.

pub fn with_level(self, level: bool) -> Self

Sets whether or not span metadata should include the tracing verbosity level information as a level field.

The level is always added to events, and based on OpenTelemetryLayer::with_error_events_to_status error-level events will mark the span status as an error.

By default, level information is disabled.