Struct tracing_opentelemetry::OpenTelemetryLayer

pub struct OpenTelemetryLayer<S, T> { /* private fields */ }

An OpenTelemetry propagation layer for use in a project that uses tracing.

Implementations

impl<S, T> OpenTelemetryLayer<S, T> where
S: Subscriber + for<'span> LookupSpan<'span>,
T: Tracer + PreSampledTracer + 'static,

pub fn new(tracer: T) -> Self

Set the Tracer that this layer will use to produce and track OpenTelemetry Spans.

Examples

use tracing_opentelemetry::OpenTelemetryLayer;
use tracing_subscriber::layer::SubscriberExt;
use tracing_subscriber::Registry;

// Create a jaeger exporter pipeline for a `trace_demo` service.
let tracer = opentelemetry_jaeger::new_pipeline()
    .with_service_name("trace_demo")
    .install_simple()
    .expect("Error initializing Jaeger exporter");

// Create a layer with the configured tracer
let otel_layer = OpenTelemetryLayer::new(tracer);

// Use the tracing subscriber `Registry`, or any other subscriber
// that impls `LookupSpan`
let subscriber = Registry::default().with(otel_layer);

pub fn with_tracer<Tracer>(
self,
tracer: Tracer
) -> OpenTelemetryLayer<S, Tracer> where
Tracer: Tracer + PreSampledTracer + 'static,

Set the Tracer that this layer will use to produce and track OpenTelemetry Spans.

Examples

use tracing_subscriber::layer::SubscriberExt;
use tracing_subscriber::Registry;

// Create a jaeger exporter pipeline for a `trace_demo` service.
let tracer = opentelemetry_jaeger::new_pipeline()
    .with_service_name("trace_demo")
    .install_simple()
    .expect("Error initializing Jaeger exporter");

// Create a layer with the configured tracer
let otel_layer = tracing_opentelemetry::layer().with_tracer(tracer);

// Use the tracing subscriber `Registry`, or any other subscriber
// that impls `LookupSpan`
let subscriber = Registry::default().with(otel_layer);

pub fn with_exception_fields(self, exception_fields: bool) -> Self

Sets whether or not span and event metadata should include OpenTelemetry exception fields such as exception.message and exception.backtrace when an Error value is recorded. If multiple error values are recorded on the same span/event, only the most recently recorded error value will show up under these fields.

These attributes follow the OpenTelemetry semantic conventions for exceptions.

By default, these attributes are not recorded.

pub fn with_exception_field_propagation(
self,
exception_field_propagation: bool
) -> Self

Sets whether or not reporting an Error value on an event will propagate the OpenTelemetry exception fields such as exception.message and exception.backtrace to the corresponding span. You do not need to enable with_exception_fields in order to enable this. If multiple error values are recorded on the same span/event, only the most recently recorded error value will show up under these fields.

These attributes follow the OpenTelemetry semantic conventions for exceptions.

By default, these attributes are not propagated to the span.

pub fn with_location(self, location: bool) -> Self

Sets whether or not span and event metadata should include OpenTelemetry attributes with location information, such as the file, module and line number.

These attributes follow the OpenTelemetry semantic conventions for source locations.

By default, locations are enabled.

pub fn with_event_location(self, event_location: bool) -> Self

👎 Deprecated since 0.17.3:
renamed to OpenTelemetrySubscriber::with_location

Sets whether or not span and event metadata should include OpenTelemetry attributes with location information, such as the file, module and line number.

These attributes follow the OpenTelemetry semantic conventions for source locations.

By default, locations are enabled.

pub fn with_tracked_inactivity(self, tracked_inactivity: bool) -> Self

Sets whether or not spans metadata should include the busy time (total time for which it was entered), and idle time (total time the span existed but was not entered).

pub fn with_threads(self, threads: bool) -> Self

Sets whether or not spans record additional attributes for the thread name and thread ID of the thread they were created on, following the OpenTelemetry semantic conventions for threads.

By default, thread attributes are enabled.